azure_copilot/requirements.txt
# Web framework
fastapi
uvicorn

# Azure SDKs
# Web framework
fastapi
uvicorn

# Azure SDKs
azure-ai-ml
azure-identity
azure-search-documents
azure-ai-openai

# Power Automate (via HTTP requests)
requests

# Environment variable management
python-dotenv

# Data processing
pandas

# Optional: for async HTTP
httpx

# OpenAI API
openai
